added new function to reduce code duplication
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@8092 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tine20/Felamimail/Controller/Message.php
@@ -10,6 +10,7 @@
  * @version     $Id$
  * 
  * @todo        add acl
+ * @todo        remove obsolete code
  */
 
 /**
@@ -161,12 +162,8 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
         
         Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . ' Getting message ' . $message->subject);
         
-        $folderBackend = new Felamimail_Backend_Folder();
-        $folder = $folderBackend->get($message->folder_id);
-        
-        try {
-            $imapBackend            = Felamimail_Backend_ImapFactory::factory($folder->account_id);
-            $backendFolderValues    = $imapBackend->selectFolder($folder->globalname);
+        if ($imapBackend = $this->_getBackendAndSelectFolder($message->folder_id, $folder)) {
+            
             $imapMessage            = $imapBackend->getMessage($message->messageuid);
             
             /********* add body ****************/
@@ -181,6 +178,59 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
             }
             
             /********* add attachments *********/
+            
+            $attachments   = array();
+            $messageParts = $imapMessage->countParts();
+            if ( $messageParts > 1) {
+                Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . ' Get ' 
+                    . $messageParts-1 . ' attachments.'
+                );
+                $partNumber = 2;
+                while ($partNumber <= $messageParts) {
+                    $part = $imapMessage->getPart($partNumber);
+
+                    $attachment = $part->getHeaders();
+                    if (isset($attachment['content-disposition'])) {
+                        preg_match("/filename=\"([a-zA-Z0-9\-\._]+)\"/", $attachment['content-disposition'], $matches);
+                        
+                        $attachment['filename']     = $matches[1];
+                        $attachment['partId']       = $partNumber;
+                        $attachment['messageId']    = $message->getId();
+                        $attachment['accountId']    = $folder->account_id;
+                        $attachment['size']         = $part->getSize();
+                                            
+                        $attachments[] = $attachment; 
+                    }
+                    
+                    $partNumber++;
+                } 
+            }
+            $message->attachments = $attachments;
+            
+            //Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . ' ' . print_r($message->toArray(), true));
+            
+            // set \Seen flag
+            if (preg_match('/\\Seen/', $message->flags) === 0) {
+                Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . ' Add \Seen flag to msg uid ' . $message->messageuid);
+                $this->addFlags($message, array(Zend_Mail_Storage::FLAG_SEEN), $folder);
+            }
+        }
+        /*
+        
+        $folderBackend = new Felamimail_Backend_Folder();
+        $folder = $folderBackend->get($message->folder_id);
+        
+        try {
+            $imapBackend            = Felamimail_Backend_ImapFactory::factory($folder->account_id);
+            $backendFolderValues    = $imapBackend->selectFolder($folder->globalname);
+            $imapMessage            = $imapBackend->getMessage($message->messageuid);
+            
+            $message->body          = $imapMessage->getBody(Zend_Mime::TYPE_TEXT);
+            
+            $message->headers       = '';
+            foreach ($imapMessage->getHeaders() as $name => $value) {  
+                $message->headers  .= "<b>$name:</b> " . substr($value,0,40) . "\n";
+            }
             
             $attachments   = array();
             $messageParts = $imapMessage->countParts();
@@ -219,6 +269,7 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
             // no imap connection -> no body
             Tinebase_Core::getLogger()->warn(__METHOD__ . '::' . __LINE__ . ' ' . $zmpe->getMessage());
         }
+        */
         
         return $message;
     }
@@ -236,10 +287,21 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
         // remove from cache db table
         parent::_deleteRecord($_record);
         
-        // remove from server
+        if ($imapBackend = $this->_getBackendAndSelectFolder($_record->folder_id, $folder)) {
+            // remove from server
+            if ($folder->globalname == 'Trash') {
+                // only delete if in Trash
+                $imapBackend->removeMessage($_record->messageuid);
+            } else {
+                // move to trash
+                Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . " Moving message '" . $_record->subject . "' to Trash.");
+                $imapBackend->moveMessage($_record->messageuid, 'Trash');
+            }
+        }
+        
+        /*
         $folderBackend = new Felamimail_Backend_Folder();
         $folder = $folderBackend->get($_record->folder_id);
-        
         try {
             $imapBackend            = Felamimail_Backend_ImapFactory::factory($folder->account_id);
             if ($imapBackend->getCurrentFolder() != $folder->globalname) {
@@ -259,6 +321,7 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
             // no imap connection -> no delete on server
             Tinebase_Core::getLogger()->warn(__METHOD__ . '::' . __LINE__ . ' ' . $zmpe->getMessage());
         }
+        */
     }
     
     /************************* other public funcs *************************/
@@ -274,6 +337,7 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
     {
         Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . ' Add flags: ' . print_r($_flags, TRUE));
         
+        /*
         if ($_folder === NULL) {
             $folderBackend = new Felamimail_Backend_Folder();
             $folder = $folderBackend->get($_message->folder_id);
@@ -287,12 +351,15 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
         if($imapBackend->getCurrentFolder() != $folder->globalname) {
             $imapBackend->selectFolder($folder->globalname);
         }
+        */
         
         // save each flag in backend, cache db and message record
-        $imapBackend->addFlags($_message->messageuid, array_intersect($_flags, array_keys(self::$_allowedFlags)));
-        foreach ($_flags as $flag) {
-            $_message->flags .= ' ' . $flag;
-            $this->_backend->addFlag($_message, $flag);
+        if ($imapBackend = $this->_getBackendAndSelectFolder($_message->folder_id, $_folder)) {
+            $imapBackend->addFlags($_message->messageuid, array_intersect($_flags, array_keys(self::$_allowedFlags)));
+            foreach ($_flags as $flag) {
+                $_message->flags .= ' ' . $flag;
+                $this->_backend->addFlag($_message, $flag);
+            }
         }
     }
     
@@ -305,6 +372,7 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
      */
     public function clearFlags($_message, $_flags, $_folder = NULL)
     {
+        /*
         if ($_folder === NULL) {
             $folderBackend = new Felamimail_Backend_Folder();
             $folder = $folderBackend->get($_message->folder_id);
@@ -318,12 +386,15 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
         if($imapBackend->getCurrentFolder() != $folder->globalname) {
             $imapBackend->selectFolder($folder->globalname);
         }
+        */
         
         // remove flag in imap backend, cache db and message record
-        $imapBackend->clearFlags($_message->messageuid, $_flags);
-        foreach ($_flags as $flag) {
-            $this->_backend->clearFlag($_message->getId(), $flag);
-        }        
+        if ($imapBackend = $this->_getBackendAndSelectFolder($_message->folder_id, $_folder)) {
+            $imapBackend->clearFlags($_message->messageuid, $_flags);
+            foreach ($_flags as $flag) {
+                $this->_backend->clearFlag($_message->getId(), $flag);
+            }
+        }
     }
     
     /**
@@ -335,6 +406,7 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
      */
     public function moveMessages($_ids, $_folderId)
     {
+        /*
         $folderBackend = new Felamimail_Backend_Folder();
         $folder = $folderBackend->get($_folderId);
 
@@ -345,24 +417,27 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
             Tinebase_Core::getLogger()->warn(__METHOD__ . '::' . __LINE__ . ' ' . $zmpe->getMessage());
             return FALSE;
         }
+        */
         
-        $messages = $this->_backend->getMultiple($_ids);
-        Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . 
-            ' Moving ' . count($messages) . ' messages to folder ' . $folder->globalname);
-        
-        // select source folder
-        $firstMessage = $messages->getFirstRecord();
-        $sourceFolder = $folderBackend->get($firstMessage->folder_id);
-        if($imapBackend->getCurrentFolder() != $sourceFolder->globalname) {
-            $imapBackend->selectFolder($sourceFolder->globalname);
+        if ($imapBackend = $this->_getBackendAndSelectFolder($_folderId, $folder, FALSE)) {
+            $messages = $this->_backend->getMultiple($_ids);
+            Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . 
+                ' Moving ' . count($messages) . ' messages to folder ' . $folder->globalname);
+            
+            // select source folder
+            $firstMessage = $messages->getFirstRecord();
+            $sourceFolder = $folderBackend->get($firstMessage->folder_id);
+            if($imapBackend->getCurrentFolder() != $sourceFolder->globalname) {
+                $imapBackend->selectFolder($sourceFolder->globalname);
+            }
+            
+            foreach ($messages as $message) {
+                $imapBackend->moveMessage($message->messageuid, $folder->globalname);
+            }
+            
+            // remove from cache db table
+            $this->_backend->delete($_ids);
         }
-        
-        foreach ($messages as $message) {
-            $imapBackend->moveMessage($message->messageuid, $folder->globalname);
-        }
-        
-        // remove from cache db table
-        $this->_backend->delete($_ids);
                 
         return TRUE;
     }
@@ -430,27 +505,48 @@ class Felamimail_Controller_Message extends Tinebase_Controller_Record_Abstract
     {
         $result         = NULL;
         $message        = parent::get($_id);
-        $folderBackend  = new Felamimail_Backend_Folder();
-        $folder         = $folderBackend->get($message->folder_id);
         
-        try {
-            $imapBackend            = Felamimail_Backend_ImapFactory::factory($folder->account_id);
-            $backendFolderValues    = $imapBackend->selectFolder($folder->globalname);
+        if ($imapBackend = $this->_getBackendAndSelectFolder($message->folder_id)) {
             $imapMessage            = $imapBackend->getMessage($message->messageuid);
-            
-            //Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__ . ' Parts:' . $imapMessage->countParts());
-            
             $result                 = $imapMessage->getPart($_partId);
-            
-        } catch (Zend_Mail_Protocol_Exception $zmpe) {
-            // no imap connection -> no download
-            Tinebase_Core::getLogger()->warn(__METHOD__ . '::' . __LINE__ . ' ' . $zmpe->getMessage());
         }
         
         return $result;
     }
     
     /************************* protected funcs *************************/
+    
+    /**
+     * get imap backend and folder (and select folder)
+     *
+     * @param string                    $_folderId
+     * @param Felamimail_Backend_Folder &$_folder
+     * @param array                     &$backendFolderValues
+     * @param boolean                   $_select
+     * @return NULL|Felamimail_Backend_Imap
+     */
+    protected function _getBackendAndSelectFolder($_folderId = NULL, &$_folder = NULL, $_select = TRUE)
+    {
+        $imapBackend = NULL;
+        
+        if ($_folder === NULL || empty($_folder)) {
+            $folderBackend  = new Felamimail_Backend_Folder();
+            $_folder = $folderBackend->get($_folderId);
+        }
+        
+        try {
+            $imapBackend = Felamimail_Backend_ImapFactory::factory($_folder->account_id);
+            if ($_select && $imapBackend->getCurrentFolder() != $_folder->globalname) {
+                $backendFolderValues = $imapBackend->selectFolder($_folder->globalname);
+            }
+            
+        } catch (Zend_Mail_Protocol_Exception $zmpe) {
+            // no imap connection
+            Tinebase_Core::getLogger()->warn(__METHOD__ . '::' . __LINE__ . ' ' . $zmpe->getMessage());
+        }
+        
+        return $imapBackend;
+    }
     
     /**
      * extract values from folder filter
